requirements.txt
librosa
spleeter
fastapi
uvicorn
python-multipart
mutagen
google-cloud-speech
speech
numpy
scipy
pandas